One exclusive decimal amount parser function among different functions
