+++ ee/plugins/bitstamp/mappers/amounts_test.go
@@ -3,6 +3,9 @@ package mappers
 import (
 	"math/big"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 var testCurrencies = map[string]int{
@@ -30,15 +33,15 @@ func TestNormalizeCurrency(t *testing.T) {
 func TestIsZeroAmount(t *testing.T) {
 	t.Parallel()
 	cases := map[string]bool{
-		"":               true,
-		"0":              true,
-		"0.0":            true,
-		"0.00000000":     true,
-		"-0":             true,
-		"not-a-number":   true, // unparseable → treated as zero
-		"0.0000001":      false,
-		"1":              false,
-		"-5.00":          false,
+		"":             true,
+		"0":            true,
+		"0.0":          true,
+		"0.00000000":   true,
+		"-0":           true,
+		"not-a-number": true, // unparseable → treated as zero
+		"0.0000001":    false,
+		"1":            false,
+		"-5.00":        false,
 	}
 	for in, want := range cases {
 		if got := IsZeroAmount(in); got != want {
@@ -58,6 +61,90 @@ func TestAbsAmount(t *testing.T) {
 		if got := AbsAmount(in); got != want {
 			t.Errorf("abs %q = %q, want %q", in, got, want)
 		}
+	}
+}
+
+func TestParseDecimalAmount(t *testing.T) {
+	t.Parallel()
+	tests := []struct {
+		name      string
+		value     string
+		precision int
+		want      *big.Int
+		wantErr   bool
+	}{
+		{
+			name:      "empty string",
+			value:     "",
+			precision: 2,
+			wantErr:   true,
+		},
+		{
+			name:      "integer no decimal",
+			value:     "100",
+			precision: 2,
+			want:      big.NewInt(10000),
+		},
+		{
+			name:      "exact precision",
+			value:     "1.23",
+			precision: 2,
+			want:      big.NewInt(123),
+		},
+		{
+			name:      "fewer decimals than precision",
+			value:     "1.2",
+			precision: 2,
+			want:      big.NewInt(120),
+		},
+		{
+			name:      "trailing zeros beyond precision truncated",
+			value:     "1.23000",
+			precision: 2,
+			want:      big.NewInt(123),
+		},
+		{
+			name:      "single trailing zero beyond precision",
+			value:     "0.500",
+			precision: 2,
+			want:      big.NewInt(50),
+		},
+		{
+			name:      "non-zero digits beyond precision errors",
+			value:     "1.235",
+			precision: 2,
+			wantErr:   true,
+		},
+		{
+			name:      "precision zero with decimal errors",
+			value:     "1.0",
+			precision: 0,
+			wantErr:   true,
+		},
+		{
+			name:      "zero amount",
+			value:     "0.00",
+			precision: 2,
+			want:      big.NewInt(0),
+		},
+		{
+			name:      "large amount trailing zeros",
+			value:     "12345.678900",
+			precision: 4,
+			want:      big.NewInt(123456789),
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			got, err := ParseDecimalAmount(tc.value, tc.precision)
+			if tc.wantErr {
+				require.Error(t, err)
+				return
+			}
+			require.NoError(t, err)
+			assert.Equal(t, tc.want, got)
+		})
 	}
 }
 
@@ -124,13 +211,13 @@ func TestResolveSinglePaymentAsset(t *testing.T) {
 func TestResolveTwoAssetConversion(t *testing.T) {
 	t.Parallel()
 	cases := []struct {
-		name        string
-		amounts     map[string]string
-		wantSrcSym  string
-		wantSrcAmt  int64
-		wantDstSym  string
-		wantDstAmt  int64
-		wantOk      bool
+		name       string
+		amounts    map[string]string
+		wantSrcSym string
+		wantSrcAmt int64
+		wantDstSym string
+		wantDstAmt int64
+		wantOk     bool
 	}{
 		{
 			name: "Quentin #679 EUR -> USDC fixture",
